fix: animation hitches and animation cycle offsets
The cycle vals were only going from 1 to 31 which caused the bg
animation hitch and other issues. Turns out instead of an offset we can
just set a cycle-start and achieve the same effect as long as the
randomness is in when the sprite is spawned. This can work for
explosions.